third_party/blink/web_tests/platform/generic/fast/dom/coreDOM-element-attribute-js-null-expected.txt

This test setting various attributes of a elements to JavaScript null.

TEST SUCCEEDED: The value was the string 'null'. [tested Attr.value]
TEST SUCCEEDED: The value was the empty string. [tested Attr.nodeValue]
TEST SUCCEEDED: The value was the empty string. [tested Attr.textContent]

TEST SUCCEEDED: The value was the empty string. [tested ProcessingInstruction.data]

TEST SUCCEEDED: The value was the empty string. [tested CharacterData.data]
